requirements.txt
pandas>=1.3.0
psycopg2-binary>=2.9.0
PyQt5>=5.15.0
PyQtWebEngine>=5.15.0
folium>=0.12.0
requests>=2.26.0
pyzipper>=0.3.5
geopandas>=0.10.0
shapely>=1.8.0